+++ src/bokeh/io/browser.py
